Changed: Replace public builder API with ModelCatalog::build
Simplify catalog construction by exposing a single build entry point and moving builder mechanics to internal implementation details.

Changes:
- Removed public `ModelCatalogBuilder` exports and builder-based construction surface
- Added `ModelCatalog::build` as the only public catalog construction API
- Moved build logic into `models/catalog/internal/builder.rs` using top-level internal functions
- Updated internal wiring, docs, tests, and benchmark call sites to the new API
- Shrunk final hash tables during catalog finalization and removed redundant reserve pass

Benefits:
- Makes the public API smaller and easier to use
- Keeps construction internals private and easier to evolve
- Improves long-lived catalog memory efficiency and removes duplicate capacity work
