--- src/llm-coding-tools-core/src/models/catalog/builder.rs
@@ -1,611 +1,0 @@
-use crate::models::catalog::internal::{
-    hash_model_key, hash_provider_key, hash_state_for_seed, model_table_entry_hash,
-    provider_table_entry_hash, ModelConfigEntry, PackedEnvRange, PackedModelEntry,
-    PackedModelTableEntry, PackedProviderTableEntry, MAX_INPUT_TOKENS, MAX_MODEL_CONFIG_COUNT,
-    MAX_OUTPUT_TOKENS, MAX_PROVIDER_COUNT,
-};
-use crate::models::catalog::public::builder_types::{
-    LookupTableKind, ModelCatalogBuildError, ModelInfo, ModelSourceRow, ProviderSourceRow,
-};
-use crate::models::catalog::public::ProviderIdx;
-use crate::models::catalog::ModelCatalog;
-use crate::models::ProviderType;
-use ahash::AHashMap;
-use hashbrown::{hash_table::Entry as TableEntry, HashTable};
-use lite_strtab::{Global, StringTable, StringTableBuilder};
-use std::collections::hash_map::Entry as MapEntry;
-
-/// Maximum hash seed value.
-///
-/// This is the upper limit for reseeding attempts when hash collisions occur.
-/// Using u8::MAX allows for 256 different hash seeds (0-255).
-pub const MAX_SEED: u8 = u8::MAX;
-
-#[derive(Debug, Clone, Copy)]
-struct ProviderSourceStats {
-    provider_count: usize,
-    total_api_url_bytes: usize,
-    total_env_keys: usize,
-    total_env_key_bytes: usize,
-}
-
-/// Batch builder for [`ModelCatalog`].
-///
-/// The builder is optimized for one-shot construction from parsed source rows.
-/// It automatically reseeds and retries on hash collisions.
-#[derive(Debug, Clone)]
-pub struct ModelCatalogBuilder {
-    seed: u8,
-    hash_state: ahash::RandomState,
-    provider_table: HashTable<PackedProviderTableEntry>,
-    model_table: HashTable<PackedModelTableEntry>,
-    provider_env_ranges: Vec<PackedEnvRange>,
-    provider_entries: Vec<ProviderType>,
-    model_entries: Vec<PackedModelEntry>,
-    model_config_entries: Vec<ModelConfigEntry>,
-    model_entry_intern: AHashMap<(PackedModelEntry, ModelConfigEntry), u16>,
-    has_any_model_config: bool,
-}
-
-impl ModelCatalogBuilder {
-    /// Creates a builder with no preallocated capacity.
-    ///
-    /// # Returns
-    ///
-    /// A new empty [`ModelCatalogBuilder`].
-    #[inline]
-    pub fn new() -> Self {
-        Self::with_capacity(0, 0)
-    }
-
-    /// Creates a builder with preallocated provider and model key capacity.
-    ///
-    /// # Parameters
-    ///
-    /// * `provider_capacity` - Expected number of provider rows.
-    /// * `model_capacity` - Expected number of model rows.
-    ///
-    /// # Returns
-    ///
-    /// A new [`ModelCatalogBuilder`] with preallocated internal storage.
-    #[inline]
-    pub fn with_capacity(provider_capacity: usize, model_capacity: usize) -> Self {
-        Self {
-            seed: 0,
-            hash_state: hash_state_for_seed(0),
-            provider_table: HashTable::with_capacity(provider_capacity),
-            model_table: HashTable::with_capacity(model_capacity),
-            provider_env_ranges: Vec::with_capacity(provider_capacity),
-            provider_entries: Vec::with_capacity(provider_capacity),
-            model_entries: Vec::with_capacity(model_capacity),
-            model_config_entries: Vec::with_capacity(model_capacity),
-            model_entry_intern: AHashMap::with_capacity(model_capacity),
-            has_any_model_config: false,
-        }
-    }
-
-    /// Builds a catalog from provider and model source rows.
-    ///
-    /// This method retries internally with new hash seeds when collisions are
-    /// encountered.
-    ///
-    /// # Parameters
-    ///
-    /// * `providers` - [`ProviderSourceRow`] values keyed by provider identifier.
-    /// * `models` - [`ModelSourceRow`] values keyed by model identifier.
-    ///
-    /// # Returns
-    ///
-    /// A fully built [`ModelCatalog`] when construction succeeds.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`ModelCatalogBuildError`] when:
-    /// - input exceeds supported numeric limits,
-    /// - token limits cannot be represented in packed model entries,
-    /// - or all seed-retry attempts still result in collisions.
-    #[inline]
-    pub fn build_from_source(
-        mut self,
-        providers: &[ProviderSourceRow],
-        models: &[ModelSourceRow],
-    ) -> Result<ModelCatalog, ModelCatalogBuildError> {
-        let provider_stats = analyze_provider_rows(providers)?;
-
-        self.seed = 0;
-        self.hash_state = hash_state_for_seed(0);
-        self.clear_entries();
-        self.reserve_for_source(provider_stats.provider_count, models.len());
-
-        loop {
-            match self.populate_tables_once(providers, models) {
-                Ok(()) => break,
-                Err(ModelCatalogBuildError::HashCollision { .. }) => {
-                    self.advance_seed_and_clear()?;
-                }
-                Err(err) => return Err(err),
-            }
-        }
-
-        Ok(self.finish_with_source(providers, provider_stats))
-    }
-
-    #[inline]
-    fn reserve_for_source(&mut self, provider_count: usize, model_count: usize) {
-        self.provider_table
-            .reserve(provider_count, provider_table_entry_hash);
-        self.model_table
-            .reserve(model_count, model_table_entry_hash);
-        self.provider_env_ranges.reserve(provider_count);
-        self.provider_entries.reserve(provider_count);
-        self.model_entries.reserve(model_count);
-        self.model_config_entries.reserve(model_count);
-        self.model_entry_intern.reserve(model_count);
-    }
-
-    #[inline]
-    fn populate_tables_once(
-        &mut self,
-        providers: &[ProviderSourceRow],
-        models: &[ModelSourceRow],
-    ) -> Result<(), ModelCatalogBuildError> {
-        let mut env_start: u16 = 0;
-
-        for provider_row in providers {
-            let provider_info = &provider_row.provider;
-            let env_count = provider_info.env_vars.len() as u8;
-            self.insert_provider_row(
-                &provider_row.provider_key,
-                env_start,
-                env_count,
-                provider_info.api_type,
-            )?;
-            env_start = env_start.wrapping_add(u16::from(env_count));
-        }
-
-        for model_row in models {
-            self.insert_model_row(&model_row.model_key, model_row.model)?;
-        }
-
-        Ok(())
-    }
-
-    #[inline]
-    fn insert_provider_row(
-        &mut self,
-        provider_key: &str,
-        env_start: u16,
-        env_count: u8,
-        api_type: ProviderType,
-    ) -> Result<(), ModelCatalogBuildError> {
-        let key = hash_provider_key(&self.hash_state, provider_key);
-        let hash48 = PackedProviderTableEntry::truncate_hash48(key.as_u64());
-        let provider_idx = self.provider_entries.len() as u16;
-
-        match self.provider_table.entry(
-            hash48,
-            |existing: &PackedProviderTableEntry| existing.hash48() == hash48,
-            provider_table_entry_hash,
-        ) {
-            TableEntry::Occupied(_) => {
-                return Err(ModelCatalogBuildError::HashCollision {
-                    table: LookupTableKind::Provider,
-                    seed: self.seed,
-                });
-            }
-            TableEntry::Vacant(vacant) => {
-                vacant.insert(PackedProviderTableEntry::from_parts(
-                    key.as_u64(),
-                    provider_idx,
-                ));
-            }
-        }
-
-        self.provider_env_ranges
-            .push(PackedEnvRange::from_parts(env_start, env_count));
-        self.provider_entries.push(api_type);
-
-        Ok(())
-    }
-
-    #[inline]
-    fn insert_model_row(
-        &mut self,
-        model_key: &str,
-        info: ModelInfo,
-    ) -> Result<(), ModelCatalogBuildError> {
-        if info.max_output > MAX_OUTPUT_TOKENS {
-            return Err(ModelCatalogBuildError::MaxOutputTokensOutOfRange {
-                max_output: info.max_output,
-                max: MAX_OUTPUT_TOKENS,
-            });
-        }
-
-        if info.max_input > MAX_INPUT_TOKENS {
-            return Err(ModelCatalogBuildError::MaxInputTokensOutOfRange {
-                max_input: info.max_input,
-                max: MAX_INPUT_TOKENS,
-            });
-        }
-
-        let model_entry = PackedModelEntry::from_model_info(info);
-        let config_entry = ModelConfigEntry::from_sampling(info.temperature, info.top_p);
-        self.has_any_model_config |= !config_entry.is_none();
-
-        let model_config_idx = match self.model_entry_intern.entry((model_entry, config_entry)) {
-            MapEntry::Occupied(existing) => *existing.get(),
-            MapEntry::Vacant(vacant) => {
-                if self.model_entries.len() >= MAX_MODEL_CONFIG_COUNT {
-                    return Err(ModelCatalogBuildError::TooManyModelConfigurations {
-                        count: self.model_entries.len() + 1,
-                        max: MAX_MODEL_CONFIG_COUNT,
-                    });
-                }
-
-                let next_idx = self.model_entries.len() as u16;
-                self.model_entries.push(model_entry);
-                self.model_config_entries.push(config_entry);
-                vacant.insert(next_idx);
-                next_idx
-            }
-        };
-
-        let key = hash_model_key(&self.hash_state, model_key);
-        let hash48 = PackedModelTableEntry::truncate_hash48(key.as_u64());
-        match self.model_table.entry(
-            hash48,
-            |existing: &PackedModelTableEntry| existing.hash48() == hash48,
-            model_table_entry_hash,
-        ) {
-            TableEntry::Occupied(_) => {
-                return Err(ModelCatalogBuildError::HashCollision {
-                    table: LookupTableKind::Model,
-                    seed: self.seed,
-                });
-            }
-            TableEntry::Vacant(vacant) => {
-                vacant.insert(PackedModelTableEntry::from_parts(
-                    key.as_u64(),
-                    model_config_idx,
-                ));
-            }
-        }
-
-        Ok(())
-    }
-
-    #[inline]
-    fn advance_seed_and_clear(&mut self) -> Result<(), ModelCatalogBuildError> {
-        if self.seed == MAX_SEED {
-            return Err(ModelCatalogBuildError::HashCollisionExhausted {
-                attempts: MAX_SEED.into(),
-            });
-        }
-
-        self.seed += 1;
-        self.hash_state = hash_state_for_seed(self.seed);
-        self.clear_entries();
-        Ok(())
-    }
-
-    #[inline]
-    fn clear_entries(&mut self) {
-        self.provider_table.clear();
-        self.model_table.clear();
-        self.provider_env_ranges.clear();
-        self.provider_entries.clear();
-        self.model_entries.clear();
-        self.model_config_entries.clear();
-        self.model_entry_intern.clear();
-        self.has_any_model_config = false;
-    }
-
-    #[inline]
-    fn finish_with_source(
-        self,
-        providers: &[ProviderSourceRow],
-        provider_stats: ProviderSourceStats,
-    ) -> ModelCatalog {
-        let model_config_entries = if self.has_any_model_config {
-            Some(self.model_config_entries.into_boxed_slice())
-        } else {
-            None
-        };
-
-        ModelCatalog {
-            hash_state: self.hash_state,
-            provider_table: self.provider_table,
-            model_table: self.model_table,
-            provider_api_urls: build_provider_api_url_table(providers, provider_stats),
-            provider_env_keys: build_provider_env_key_table(providers, provider_stats),
-            provider_env_ranges: self.provider_env_ranges.into_boxed_slice(),
-            provider_entries: self.provider_entries.into_boxed_slice(),
-            model_entries: self.model_entries.into_boxed_slice(),
-            model_config_entries,
-        }
-    }
-}
-
-#[inline]
-fn analyze_provider_rows(
-    providers: &[ProviderSourceRow],
-) -> Result<ProviderSourceStats, ModelCatalogBuildError> {
-    use crate::models::catalog::internal::MAX_ENV_RANGE_COUNT;
-
-    let provider_count = providers.len();
-    if provider_count > MAX_PROVIDER_COUNT {
-        return Err(ModelCatalogBuildError::TooManyProviders {
-            count: provider_count,
-            max: MAX_PROVIDER_COUNT,
-        });
-    }
-
-    let mut total_api_url_bytes = 0usize;
-    let mut total_env_keys = 0usize;
-    let mut total_env_key_bytes = 0usize;
-
-    for provider_row in providers {
-        let provider_info = &provider_row.provider;
-        let env_count = provider_info.env_vars.len();
-        if env_count > usize::from(MAX_ENV_RANGE_COUNT) {
-            return Err(
-                ModelCatalogBuildError::TooManyProviderEnvVarsForOneProvider {
-                    count: env_count,
-                    max: usize::from(MAX_ENV_RANGE_COUNT),
-                },
-            );
-        }
-
-        total_api_url_bytes += provider_info.api_url.len();
-        total_env_keys += env_count;
-        for env_key in &provider_info.env_vars {
-            total_env_key_bytes += env_key.len();
-        }
-    }
-
-    Ok(ProviderSourceStats {
-        provider_count,
-        total_api_url_bytes,
-        total_env_keys,
-        total_env_key_bytes,
-    })
-}
-
-#[inline]
-fn build_provider_api_url_table(
-    providers: &[ProviderSourceRow],
-    stats: ProviderSourceStats,
-) -> StringTable<u32, ProviderIdx> {
-    let mut builder = StringTableBuilder::<u32, ProviderIdx>::with_capacity_in(
-        stats.provider_count,
-        stats.total_api_url_bytes,
-        Global,
-    );
-
-    for provider_row in providers {
-        builder
-            .try_push(&provider_row.provider.api_url)
-            .expect("string table insert");
-    }
-
-    builder.build()
-}
-
-#[inline]
-fn build_provider_env_key_table(
-    providers: &[ProviderSourceRow],
-    stats: ProviderSourceStats,
-) -> StringTable<u32, ProviderIdx> {
-    let mut builder = StringTableBuilder::<u32, ProviderIdx>::with_capacity_in(
-        stats.total_env_keys,
-        stats.total_env_key_bytes,
-        Global,
-    );
-
-    for provider_row in providers {
-        for env_key in &provider_row.provider.env_vars {
-            builder.try_push(env_key).expect("string table insert");
-        }
-    }
-
-    builder.build()
-}
-
-impl Default for ModelCatalogBuilder {
-    #[inline]
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::ModelCatalogBuilder;
-    use crate::models::catalog::{
-        Modality, ModelCatalogBuildError, ModelInfo, ModelSourceRow, ProviderInfo,
-        ProviderSourceRow,
-    };
-    use crate::models::ProviderType;
-
-    fn provider(api_url: &str, env_vars: &[&str], api_type: ProviderType) -> ProviderInfo {
-        ProviderInfo {
-            api_url: api_url.to_owned(),
-            env_vars: env_vars.iter().map(|s| s.to_string()).collect(),
-            api_type,
-        }
-    }
-
-    fn info(max_input: u32, max_output: u32) -> ModelInfo {
-        ModelInfo {
-            modalities: Modality::TEXT,
-            max_input,
-            max_output,
-            temperature: None,
-            top_p: None,
-        }
-    }
-
-    fn provider_row(provider_key: &str, provider: ProviderInfo) -> ProviderSourceRow {
-        ProviderSourceRow::new(provider_key, provider)
-    }
-
-    fn model_row(model_key: &str, model: ModelInfo) -> ModelSourceRow {
-        ModelSourceRow::new(model_key, model)
-    }
-
-    fn source_rows() -> (Vec<ProviderSourceRow>, Vec<ModelSourceRow>) {
-        (
-            vec![provider_row(
-                "alpha",
-                provider(
-                    "https://alpha.example",
-                    &["ALPHA_KEY"],
-                    ProviderType::OpenAiCompletions,
-                ),
-            )],
-            vec![model_row("m1", info(4096, 512))],
-        )
-    }
-
-    #[test]
-    fn build_from_source_builds_catalog() {
-        let (providers, models) = source_rows();
-        let catalog = ModelCatalogBuilder::with_capacity(1, 1)
-            .build_from_source(&providers, &models)
-            .expect("source build should succeed");
-
-        assert_eq!(catalog.provider_len(), 1);
-        assert_eq!(catalog.model_len(), 1);
-        assert!(catalog.lookup("alpha", "m1").is_some());
-    }
-
-    #[test]
-    fn duplicate_provider_keys_exhaust_reseed_attempts() {
-        let providers = vec![
-            provider_row(
-                "alpha",
-                provider("https://alpha.example", &["ALPHA_KEY"], ProviderType::Azure),
-            ),
-            provider_row(
-                "alpha",
-                provider("https://beta.example", &["BETA_KEY"], ProviderType::Azure),
-            ),
-        ];
-        let models = vec![model_row("m1", info(4096, 512))];
-
-        match ModelCatalogBuilder::new().build_from_source(&providers, &models) {
-            Err(err) => {
-                assert_eq!(
-                    err,
-                    ModelCatalogBuildError::HashCollisionExhausted {
-                        attempts: super::MAX_SEED.into()
-                    }
-                );
-            }
-            Ok(_) => panic!("duplicate provider key should collide for all seeds"),
-        }
-    }
-
-    #[test]
-    fn model_entries_are_deduplicated_by_info_and_config() {
-        let providers = vec![provider_row(
-            "alpha",
-            provider("https://alpha.example", &["ALPHA_KEY"], ProviderType::Azure),
-        )];
-        let models = vec![
-            model_row(
-                "m1",
-                ModelInfo {
-                    modalities: Modality::TEXT,
-                    max_input: 4096,
-                    max_output: 512,
-                    temperature: Some(1.0),
-                    top_p: Some(0.9),
-                },
-            ),
-            model_row(
-                "m2",
-                ModelInfo {
-                    modalities: Modality::TEXT,
-                    max_input: 4096,
-                    max_output: 512,
-                    temperature: Some(1.0),
-                    top_p: Some(0.9),
-                },
-            ),
-        ];
-
-        let catalog = ModelCatalogBuilder::new()
-            .build_from_source(&providers, &models)
-            .expect("source build should succeed");
-
-        assert_eq!(catalog.model_len(), 2);
-        assert_eq!(catalog.model_config_len(), 1);
-    }
-
-    #[test]
-    fn too_many_provider_env_vars_returns_error() {
-        let providers = vec![provider_row(
-            "alpha",
-            provider(
-                "https://alpha.example",
-                &["A", "B", "C", "D"],
-                ProviderType::Azure,
-            ),
-        )];
-        let models = vec![model_row("m1", info(4096, 512))];
-
-        match ModelCatalogBuilder::new().build_from_source(&providers, &models) {
-            Err(err) => {
-                assert_eq!(
-                    err,
-                    ModelCatalogBuildError::TooManyProviderEnvVarsForOneProvider {
-                        count: 4,
-                        max: 3,
-                    }
-                );
-            }
-            Ok(_) => panic!("provider with too many env vars should fail"),
-        }
-    }
-
-    #[test]
-    fn max_output_tokens_out_of_range_returns_error() {
-        let (providers, _) = source_rows();
-        let max_output = super::MAX_OUTPUT_TOKENS;
-        let models = vec![model_row("m1", info(4096, max_output.saturating_add(1)))];
-
-        match ModelCatalogBuilder::new().build_from_source(&providers, &models) {
-            Err(err) => {
-                assert_eq!(
-                    err,
-                    ModelCatalogBuildError::MaxOutputTokensOutOfRange {
-                        max_output: max_output.saturating_add(1),
-                        max: max_output,
-                    }
-                );
-            }
-            Ok(_) => panic!("max output over packed limit should fail"),
-        }
-    }
-
-    #[test]
-    fn max_input_tokens_out_of_range_returns_error() {
-        let (providers, _) = source_rows();
-        let max_input = super::MAX_INPUT_TOKENS;
-        let models = vec![model_row("m1", info(max_input.saturating_add(1), 512))];
-
-        match ModelCatalogBuilder::new().build_from_source(&providers, &models) {
-            Err(err) => {
-                assert_eq!(
-                    err,
-                    ModelCatalogBuildError::MaxInputTokensOutOfRange {
-                        max_input: max_input.saturating_add(1),
-                        max: max_input,
-                    }
-                );
-            }
-            Ok(_) => panic!("max input over packed limit should fail"),
-        }
-    }
-}
