+++ src/llm-coding-tools-core/src/models/mod.rs
@@ -5,6 +5,6 @@ mod provider_type;
 
 pub use catalog::{
     CatalogEntry, LookupTableKind, Modality, Model, ModelCatalog, ModelCatalogBuildError,
-    ModelCatalogBuilder, ModelInfo, ModelSourceRow, Provider, ProviderInfo, ProviderSourceRow,
+    ModelInfo, ModelSourceRow, Provider, ProviderInfo, ProviderSourceRow,
 };
 pub use provider_type::ProviderType;
